backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pandas>=2.2.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
psycopg2-binary==2.9.9
